+++ packages/plugins/plugin-cli/src/index.ts
@@ -1,0 +1,15 @@
+export {
+	parsePluginAddCommand,
+	type PluginAddCommand,
+	type PluginAddCommandDependencies,
+	runPluginAddCommand,
+	runPluginCli,
+} from "./command.js";
+export {
+	type NpmPackResult,
+	type NpmPackRunner,
+	type NpmPluginPackageManifest,
+	type ResolvedNpmPluginArchive,
+	resolveNpmPluginArchive,
+	runNpmPack,
+} from "./npm-package.js";